Make `Fact`, `ExpectedFact`, and `ReportedFact` top-level classes, and reduce their accessibility. Make `ReportedFact.matches()` take `ExpectedFact`.
