+++ conformance-test-framework/src/main/java/org/jspecify/conformance/ConformanceTestRunner.java
@@ -27,7 +27,6 @@ import static java.nio.file.Files.walk;
 import static java.util.Arrays.stream;
 import static java.util.Objects.requireNonNullElse;
 import static java.util.stream.Collectors.toList;
-import static org.jspecify.conformance.ConformanceTestRunner.ExpectedFact.readExpectedFact;
 
 import com.google.common.base.Ascii;
 import com.google.common.collect.ImmutableList;
@@ -213,153 +212,5 @@ public final class ConformanceTestRunner {
                 mode));
       }
     }
-  }
-
-  /** An expected or reported fact within a test input file. */
-  public abstract static class Fact {
-
-    private final Path file;
-    private final long lineNumber;
-
-    protected Fact(Path file, long lineNumber) {
-      this.file = file;
-      this.lineNumber = lineNumber;
-    }
-
-    /** The file path relative to the test source root. */
-    public final Path getFile() {
-      return file;
-    }
-
-    /** Returns the line number of the code in the source file to which this fact applies. */
-    public final long getLineNumber() {
-      return lineNumber;
-    }
-
-    /** The fact text. */
-    public abstract String getFactText();
-  }
-
-  /**
-   * An assertion that the tool behaves in a way consistent with a specific fact about a line in the
-   * source code. Some of these facts indicate that according to the JSpecify specification, the
-   * code in question may have an error that should be reported to users; other expected facts are
-   * informational, such as the expected nullness-augmented type of an expression.
-   */
-  public static final class ExpectedFact extends Fact {
-    private static final Pattern NULLNESS_MISMATCH =
-        Pattern.compile("jspecify_nullness_mismatch\\b.*");
-
-    private static final ImmutableList<Pattern> ASSERTION_PATTERNS =
-        ImmutableList.of(
-            NULLNESS_MISMATCH,
-            // TODO: wildcard types have whitespace
-            Pattern.compile("test:cannot-convert:\\S+ to \\S+"),
-            Pattern.compile("test:expression-type:[^:]+:.*"),
-            Pattern.compile("test:irrelevant-annotation:\\S+"),
-            Pattern.compile("test:sink-type:[^:]+:.*"));
-
-    /**
-     * Returns an expected fact representing that the source type cannot be converted to the sink
-     * type in any world.
-     */
-    public static String cannotConvert(String sourceType, String sinkType) {
-      return String.format("test:cannot-convert:%s to %s", sourceType, sinkType);
-    }
-
-    /** Returns an expected fact representing an expected expression type. */
-    public static String expressionType(String expressionType, String expression) {
-      return String.format("test:expression-type:%s:%s", expressionType, expression);
-    }
-
-    /** Returns an expected fact representing that an annotation is not relevant. */
-    public static String irrelevantAnnotation(String annotationType) {
-      return String.format("test:irrelevant-annotation:%s", annotationType);
-    }
-
-    /** Returns an expected fact representing that an annotation is not relevant. */
-    public static String sinkType(String sinkType, String sink) {
-      return String.format("test:sink-type:%s:%s", sinkType, sink);
-    }
-
-    /**
-     * Returns {@code true} if {@code fact} is a legacy {@code jspecify_nullness_mismatch}
-     * assertion.
-     */
-    public static boolean isNullnessMismatch(String fact) {
-      return NULLNESS_MISMATCH.matcher(fact).matches();
-    }
-
-    /** Read an expected fact from a line of either a source file or a report. */
-    static @Nullable String readExpectedFact(String text) {
-      return ASSERTION_PATTERNS.stream().anyMatch(pattern -> pattern.matcher(text).matches())
-          ? text
-          : null;
-    }
-
-    private final String fact;
-    private final long factLineNumber;
-
-    ExpectedFact(Path file, long lineNumber, String fact, long factLineNumber) {
-      super(file, lineNumber);
-      this.fact = fact;
-      this.factLineNumber = factLineNumber;
-    }
-
-    @Override
-    public String getFactText() {
-      return fact;
-    }
-
-    /** Returns the line number in the input file where the expected fact is. */
-    public long getFactLineNumber() {
-      return factLineNumber;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (!(o instanceof ExpectedFact)) {
-        return false;
-      }
-      ExpectedFact that = (ExpectedFact) o;
-      return this.getFile().equals(that.getFile())
-          && this.getLineNumber() == that.getLineNumber()
-          && this.fact.equals(that.fact);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(getFile(), getLineNumber(), fact);
-    }
-  }
-
-  /** A fact reported by the analysis under test. */
-  public abstract static class ReportedFact extends Fact {
-    protected ReportedFact(Path file, long lineNumber) {
-      super(file, lineNumber);
-    }
-
-    @Override
-    public String getFactText() {
-      return requireNonNullElse(expectedFact(), toString());
-    }
-
-    /** Returns true if this reported fact must match an {@link ExpectedFact}. */
-    protected abstract boolean mustBeExpected();
-
-    /** Returns true if this reported fact matches the given expected fact. */
-    protected boolean matches(String expectedFact) {
-      return expectedFact.equals(expectedFact());
-    }
-
-    /** Returns the equivalent expected fact. */
-    protected abstract @Nullable String expectedFact();
-
-    /** Returns the message reported, without the file name or line number. */
-    @Override
-    public abstract String toString();
   }
 }
